Consolidate globe specific code
